@@ -1,16 +1,10 @@
-namespace SharpDistributions
+namespace SharpDistributions;
+
+public class ProbabilitySetElement<T> : IProbabilitySetElement<T>
 {
-    public interface IProbabilitySetElement<T>
-    {
-        Distribution<T> Distribution { get; set; }
-    }
-
-    public class ProbabilitySetElement<T> : IProbabilitySetElement<T>
-    {
-        public T Element { get; set; }
+    public T Element { get; set; }
 
 
-        public Distribution<T> Distribution { get; set; }
+    public Distribution<T> Distribution { get; set; }
 
-    }
 }
