@@ -1,32 +1,57 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using SharpDistributions;
+using SharpDistributions.Sample01;
+
+// Use directly of Sampling Functions
+// We don't need Distribution extra facilities;
+var b = Distributions.Bernoulli(0.3);
+var count = 0;
+for (var i = 0; i < 100000; i++)
+{
+    if (b.NextSample()) count++;
+}
+Console.WriteLine(count);
+
+var a = Army.ArmyElements().Take(20000).ToList();
+            
+var sa = a.OfType<SoldierTypeA>().Count();
+var sb = a.OfType<SoldierTypeB>().Count();
+var ta = a.OfType<TankA>().Count();
+var tb = a.OfType<TankB>().Count();
+           
+Console.WriteLine($"Soldiers/Tanks {a.OfType<Soldier>().Count()/ (double)a.OfType<Tank>().Count()}");
+Console.WriteLine($"Soldiers A({sa}) B({sb}), Tanks A({ta}) B({tb})");
 
 namespace SharpDistributions.Sample01
 {
-    interface IArmyElement
+    internal interface IArmyElement
     {
         
     }
-    class Tank : IArmyElement { }
-    class Soldier : IArmyElement { }
 
-    class SoldierTypeA : Soldier
-    {
-    }
-    class SoldierTypeB : Soldier
-    {
-    }
+    internal class Tank : IArmyElement { }
 
-    class TankA : Tank
+    internal class Soldier : IArmyElement { }
+
+    internal class SoldierTypeA : Soldier
     {
     }
 
-    class TankB : Tank
+    internal class SoldierTypeB : Soldier
     {
     }
 
-    class Army
+    internal class TankA : Tank
+    {
+    }
+
+    internal class TankB : Tank
+    {
+    }
+
+    internal class Army
     {
         public List<Tank> Tanks;
         public List<Soldier> Soldiers;
@@ -82,31 +107,6 @@ namespace SharpDistributions.Sample01
         static IEnumerator<Soldier> SoldierSamplingFunction()
         {
             return Distributions.Bernoulli(0.6).Select(b => b ? new SoldierTypeA() : new SoldierTypeB() as Soldier).GetEnumerator();
-        }
-    }
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            // Use directly of Sampling Functions
-            // We don't need Distribution extra facilities;
-            var b = Distributions.Bernoulli(0.3);
-            var count = 0;
-            for (var i = 0; i < 100000; i++)
-            {
-                if (b.NextSample()) count++;
-            }
-            Console.WriteLine(count);
-
-            var a = Army.ArmyElements().Take(20000).ToList();
-            
-            var sa = a.OfType<SoldierTypeA>().Count();
-            var sb = a.OfType<SoldierTypeB>().Count();
-            var ta = a.OfType<TankA>().Count();
-            var tb = a.OfType<TankB>().Count();
-           
-            Console.WriteLine($"Soldiers/Tanks {a.OfType<Soldier>().Count()/ (double)(a.OfType<Tank>().Count())}");
-            Console.WriteLine($"Soldiers A({sa}) B({sb}), Tanks A({ta}) B({tb})");
         }
     }
 }
